# back/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Cette API a pour but d'être le back end du projet Creative Beats",
    "version": "1.0.0",
    "title": "Swagger Creative Beats",
    "contact": {
      "email": "jonathan.ramassamy@u-psud.fr"
    }
  },
...
  "paths": {
...
    "/userAPI/addMusic": {
      "post": {
        "tags": [
          "/userAPI/"
        ],
        "summary": "Ajouter une musique à un utilisateur",
        "operationId": "addMusic",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userTag": {
                    "type": "string"
                  },
                  "userMusic": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/userAPI/loadUserByTag": {
      "get": {
        "tags": [
          "/userAPI/"
        ],
        "summary": "Ajouter une musique à un utilisateur",
        "operationId": "loadUserByTag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userTag": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "description": "Un objet contenant une propriété userTag",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "_id",
          "email",
          "userAlias",
          "userTag",
          "userPwd",
          "userRole",
          "userPoints",
          "userDescription",
          "userLinks",
          "musics",
          "userVerified",
          "telCrypted"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "userAlias": {
            "type": "string"
          },
          "userTag": {
            "type": "string"
          },
          "userPwd": {
            "type": "string",
            "format": "password"
          },
          "userRole": {
            "type": "string"
          },
          "userPoints": {
            "type": "string"
          },
          "userDescription": {
            "type": "string"
          },
          "userLinks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "musics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "userVerified": {
            "type": "boolean"
          },
          "telCrypted": {
            "type": "string"
          }
        }
      },
...
    }
  }
}